Profile half done with profile controller

+++ resources/views/events/event_by_id.blade.php
@@ -35,23 +35,25 @@
                         <br>
                         <span class="float-left"> <b>Location:</b> {{$event->location->address}}</span>
                         <br><br>
-
-                        @if(Auth::check())
-                            @if(!$event->attending)
-                                {!! Form::open(['method' => 'POST', 'route' => ['attend.store']]) !!}
-                                {{ method_field('POST') }}
-                                    <input type="hidden" name="event_id" value="{{ $event->id }}">
-                                    <input type="hidden" name="attending" value="1">
-                                    <button type="submit" class="btn btn-sm btn-primary">Attend</button>
-                            @else
+                        {{-- @if(Auth::check())
+                        @if($user_attending != null)
+                            @if($user_attending[0]['user_id'] == Auth::user()->id && $user_attending[0]['event_id'] == Request::segment(2))
                                 {{Form::open(['method'  => 'DELETE', 'route' => ['attend.destroy', $event->attending->id]])}}
-                                    <button type="submit" class="btn btn-sm btn-warning">Don't Attend</button>
+                                <button type="submit" class="btn btn-sm btn-warning">Don't Attend</button>
                             @endif
-                                @csrf
-                                <button type="button" class="btn btn-success btn-sm" data-toggle="modal" data-target="#myModal{{ $event->id }}">Map</button>
-                                <button type="button" class="btn btn-info btn-sm float-right" data-toggle="modal" data-target="#attending">See Attending {{$attending}}</button>
-                            {!! Form::close() !!}
                         @endif
+                        @if(empty($user_attending))
+                            {!! Form::open(['method' => 'POST', 'route' => ['attend.store']]) !!}
+                            {{ method_field('POST') }}
+                            <input type="hidden" name="event_id" value="{{ $event->id }}">
+                            <input type="hidden" name="attending" value="1">
+                            <button type="submit" class="btn btn-sm btn-primary">Attend</button>
+                        @endif
+                                @csrf
+                                --}}<button type="button" class="btn btn-success btn-sm" data-toggle="modal" data-target="#myModal{{ $event->id }}">Map</button>{{--
+                                <button type="button" class="btn btn-info btn-sm float-right" data-toggle="modal" data-target="#attending">See Attending {{$attend_count}}</button>
+                            {!! Form::close() !!}
+                        @endif --}}
 
 
                         {{-- List of people attending this event --}}
@@ -63,9 +65,7 @@
                                 Attending
                             @endslot
                             @slot('body')
-                                @foreach($event->users_attend as $user_attend)
-                                    {{$user_attend->name}}
-                                @endforeach
+                                NAMES
                             @endslot
                         @endcomponent
                         {{-- List of people attending this event END--}}
@@ -110,7 +110,7 @@
                                     });
                                   }
                                 </script>
-                                <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyACKJBGE39ESFPS3fCDiY7hH9MdPfIf6CA&callback=initMap">
+                                <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyB58LtIzgABo3BfoTaI4XL4gMNmx0XfSnA&callback=initMap">
                                 </script>
 
                             @endslot
